Show new player groups after sync selection.
LMS only send player status updates for the player being synced or
unsynced, not for other accected players, even if status updates are
subscribed.

To handle this we request a single player status for all players after
we send a player sync command.

No default selection in the player sync dialog and disable the Ok button
until a selection is made.

While at it upgrade the player sync dialog source code, and remove
warnings.